Provide nvfancontrol as a service for the nvpmodel snap
jetson_clocks runs the following commands
before reading the clocks and the fan status.
    systemctl is-active nvfancontrol
​    systemctl is-active nv-spi-server

Unfortunately running systemctl from a snap is not possible.
Alternatively snapctl can be used.
    https://forum.snapcraft.io/t/enable-disable-systemd-services/10626/2

So the nvidia-l4t-nvfancontrol stage package has been included,
and usr/sbin/nvfancontrol has been set as a daemon to be monitored by snapctl.

Finally a bash script is executed instead of "systemctl is-active" to run "snapctl services"
to retrieve the status of nvfancontrol.

Important notes:
1. Currently the nvfancontrol service cannot be fully started since the snap does not have access to /var/lock/nvf.lock,
   However this functionality does not impact our use case.
2. Checking the nv-spi-server service status is currently being bypassed, similarly it does not impact our use case.